% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MNB.R
\name{MNB}
\alias{MNB}
\alias{package}
\title{Diagnostic tools for a multivariate negative binomial model}
\description{
Diagnostic tools as residual analysis, global, local and total-local influence
     for the multivariate model from the random intercept Poisson-GlG mode. Including
     also, the estimation process by maximum likelihood and generating multivariate
     negative binomial data.
}
\section{MNB package functions}{

\itemize{
\item \code{\link[MNB]{rMNB}}
\item \code{\link[MNB]{fit.MNB}}
\item \code{\link[MNB]{qMNB}}
\item \code{\link[MNB]{re.MNB}}
\item \code{\link[MNB]{envelope.MNB}}
\item \code{\link[MNB]{global.MNB}}
\item \code{\link[MNB]{local.MNB}}
}
}

\references{
\itemize{
\item Fabio, L. C., Villegas, C., Carrasco, J. M. F., and de Castro, M. (2023).
Diagnostic tools for a multivariate negative binomial
model for fitting correlated data with overdispersion. Communications in
Statistics - Theory and Methods, 52, 1833–1853.
\item Fabio, L. C., Villegas, C., Mamun, A. S., and Carrasco, J. M. F. (2025).
Residual analysis for discrete correlated data in the multivariate approach.
Brazilian Journal of Biometrics, 43, e43728.
}
}
\author{
Jalmar M F Carrasco <carrascojalmar@gmail.com>,
Cristian M Villegas Lobos <master.villegas@gmail.com> and Lizandra C Fabio <lizandrafabio@gmail.com>
}
